Delete option modified

+++ src/app/user/user-details/user-details.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { Route, Router } from '@angular/router';
 import { UserService } from '@src/app/services/userServices/user.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-user-details',
@@ -26,11 +27,32 @@ export class UserDetailsComponent {
   editUser(userId: number) {
     this.router.navigate(['/user/edit', userId]);
   }
-  displayedColumns: string[] = ['name', 'email', 'dob', 'phone', 'gender', 'actions'];
+  displayedColumns: string[] = [
+    'name',
+    'email',
+    'dob',
+    'phone',
+    'gender',
+    'actions',
+  ];
 
   deleteUser(userId: number) {
-    this.userService.deleteUser(userId);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You are about to delete this user.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.userService.deleteUser(userId);
 
-    this.users = this.users.filter((user) => user.id !== userId);
+        this.users = this.users.filter((user) => user.id !== userId);
+
+        Swal.fire('Deleted!', 'The user has been deleted.', 'success');
+      }
+    });
   }
 }
